+++ src/redirection.c
@@ -22,8 +22,8 @@ void redirection_output(char **str, var_t *var, bool overwrite)
 
 void redirection_input(char **str, var_t *var)
 {
-    unsigned int i = var->indice + 2;
-    unsigned int j = var->indice;
+    size_t i = var->indice + 2;
+    size_t j = var->indice;
 
     if ((var->fd = open(str[var->indice + 1], O_RDONLY)) == -1)
         exit(EXIT_FAILURE);
